+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Uploading documents with special characters
 - Deleting entire directories
 - Over-fetching directories content even when not used
+- Over-fetching files content even when not used
 
 ## 4.0.0 - 2023-11-01